+++ App/Redux/Common/MyClasses/Selectors.js
@@ -2,7 +2,6 @@ import { createSelector } from 'reselect';
 import Lodash from 'lodash';
 
 const getMyClassesManager = objState => Lodash.get(objState, ['MyClasses'], {});
-const getMyUserManager = objState => Lodash.get(objState, ['userManager'], {});
 
 export const getMyClassesLookup = createSelector(
   getMyClassesManager,
@@ -18,16 +17,7 @@ export const getMyClasses = createSelector(
   }
 );
 
-export const isLogged = createSelector(
-  getMyUserManager,
-  objState => {
-    const value = Lodash.get(objState, ['token'], null);
-    return value != null;
-  }
-);
-
 export default {
   getMyClasses,
   getMyClassesLookup,
-  isLogged,
 };
